use uint64_t instead of int

+++ day1/day1.cpp
@@ -7,7 +7,7 @@
 
 namespace ranges = std::ranges;
 
-int count_increased(const std::vector<int> &depths)
+int count_increased(const std::vector<uint64_t> &depths)
 {
     int increased = 0;
     for (unsigned int i = 1; i < depths.size(); i++)
@@ -26,7 +26,7 @@ int main()
 
     aoc::assert_equal(count_increased(depths), 1266); // part 1
 
-    std::vector<int> three_meas;
+    std::vector<uint64_t> three_meas;
     for (unsigned int i = 2; i < depths.size(); i++)
     {
         three_meas.push_back(depths[i] + depths[i - 1] + depths[i - 2]);
